requirements.txt
absl-py==2.0.0
addict==2.4.0
aiofiles==23.2.1
albumentations==1.3.1
aliyun-python-sdk-core==2.15.2
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
apptools==5.2.1
asttokens==2.4.1
attrs==24.2.0
batchgenerators==0.25
black==24.8.0
blinker==1.8.2
cachetools==5.3.1
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==2.1.1
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cloudpickle==3.0.0
cmake==3.29.2
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
ConfigArgParse==1.7
configobj==5.0.8
contourpy==1.1.1
coverage==7.3.4
crcmod==1.7
cryptography==43.0.1
cycler==0.11.0
Cython==3.0.11
dash==2.18.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
datasketch==1.6.5
decorator==5.1.1
deepdiff==8.1.1
Deprecated==1.2.14
dipy==1.10.0
docker-pycreds==0.4.0
edt==2.4.0
efficientnet_pytorch==0.7.1
einops==0.6.1
envisage==7.0.3
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.112.2
fastjsonschema==2.20.0
ffmpy==0.4.0
filelock==3.14.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.42.1
fsspec==2023.9.2
ftfy==6.2.3
future==0.18.3
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.40.2
google-auth-oauthlib==1.0.0
grpcio==1.58.0
gviz-api==1.10.0
h11==0.14.0
h5py==3.10.0
hausdorff==0.2.6
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
humanfriendly==10.0
humanize==4.9.0
hydra-core==1.3.2
icecream==2.1.3
idna==3.4
imageio==2.31.4
importlib_metadata==8.5.0
importlib_resources==6.4.4
iopath==0.1.9
ipdb==0.13.13
ipython==8.27.0
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.5
kneed==0.8.5
kornia==0.7.1
lazy_loader==0.3
linecache2==1.0.0
littleutils==0.2.2
llvmlite==0.41.1
lxml==5.3.0
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.8.0
matplotlib-inline==0.1.7
mdurl==0.1.2
MedPy==0.4.0
medutils==0.1.21
mmengine==0.10.5
mmsegmentation==1.2.2
model-index==0.1.11
monai==1.3.0
mpmath==1.3.0
munch==4.0.0
mypy-extensions==1.0.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.0
nibabel==5.1.0
nilearn==0.10.4
numba==0.58.1
numpy==1.23.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.9.41
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.16.1
onnxruntime==1.19.0
open3d==0.18.0
opencv-python==4.8.0.76
opencv-python-headless==4.8.0.76
opendatalab==0.0.10
openmim==0.3.9
openpyxl==3.1.2
openxlab==0.1.1
ordered-set==4.1.0
orderly-set==5.2.3
orjson==3.10.7
oss2==2.17.0
outdated==0.2.2
packaging==24.1
pandas==2.1.4
pandas-flavor==0.6.0
parameterized==0.9.0
parso==0.8.4
pathspec==0.12.1
patsy==0.5.4
pexpect==4.9.0
Pillow==9.3.0
pingouin==0.5.3
platformdirs==4.2.1
plotly==5.24.1
portalocker==2.10.1
prefetch_generator==1.0.3
pretrainedmodels==0.7.4
prettytable==3.11.0
prompt_toolkit==3.0.48
protobuf==4.24.3
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.8
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.8.2
pydantic_core==2.20.1
pydicom==3.0.1
pydub==0.25.1
pyface==8.0.0
Pygments==2.17.2
pyparsing==3.1.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
pyquaternion==0.9.9
PySimpleGUI==5.0.8.3
python-dateutil==2.8.2
python-multipart==0.0.9
pytorch_ssim==0.1
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
qudida==0.0.4
read-roi==1.6.0
referencing==0.35.1
regex==2024.9.11
requests==2.28.2
requests-oauthlib==2.0.0
retrying==1.3.4
rich==13.4.2
roifile==2024.1.10
rpds-py==0.20.0
rsa==4.9
ruff==0.6.2
safetensors==0.3.3
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.11.3
seaborn==0.13.0
seg-metrics==1.1.6
segmentation-models-pytorch==0.3.3
semantic-version==2.10.0
sentry-sdk==2.29.1
setproctitle==1.3.3
setuptools-scm==8.1.0
shapely==2.0.3
shellingham==1.5.4
SimpleITK==2.3.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
stack-data==0.6.3
starlette==0.38.2
statsmodels==0.14.1
sympy==1.12
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.14.1
tensorboard-data-server==0.7.1
tensorboard-plugin-profile==2.13.1
tensorboardX==2.6.2.2
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.9.26
timm==0.9.2
tomli==2.0.1
tomlkit==0.12.0
torch==2.2.2
torch-cka==0.21
torchaudio==2.2.2
torchcam==0.4.0
torchio==0.19.6
torchsummary==1.5.1
torchvision==0.17.2
tqdm==4.65.2
traceback2==1.4.0
traitlets==5.14.3
traits==6.4.3
traitsui==8.0.0
triton==2.2.0
trx-python==0.3
tsnecuda==3.0.1
typer==0.12.3
typing_extensions==4.12.2
tzdata==2023.3
unittest2==1.1.0
urllib3==1.26.20
uvicorn==0.30.6
vtk==9.3.0
wandb==0.17.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.4
widgetsnbextension==4.0.13
wrapt==1.16.0
xarray==2023.12.0
yacs==0.1.8
yapf==0.40.2
zipp==3.20.2